+++ R/use_ai.R
@@ -51,7 +51,7 @@ use_ai <- function(
     stringr::str_replace_all(skills, "-", "_")
   )
   skill_paths <- purrr::map2(skills, skill_fn_names, function(skill, fn_name) {
-    fn <- match.fun(fn_name)
+    fn <- get(fn_name, envir = asNamespace("pkgskills"))
     extra_args <- if (skill == "create-issue") {
       list(gh_token = gh_token)
     } else {